cybergym-purple/pyproject.toml
[project]
name = "cybergym-purple-agent"
version = "0.1.0"
description = "A2A purple agent for CyberGym — analyzes vulnerabilities and generates PoC exploits"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20,<1.0.0",
    "httpx>=0.28.1",
    "openai>=1.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
